+++ internal/image/docker_schema1_test.go
