+++ internal/image/docker_schema2_test.go
