+++ internal/image/fixtures/oci1-all-media-types-to-schema2.json
