+++ internal/image/fixtures/oci1-all-media-types.json
