+++ internal/image/fixtures/oci1-config.json
