+++ internal/image/fixtures/oci1-invalid-media-type.json
