+++ internal/image/fixtures/oci1-to-schema1.json
