+++ internal/image/fixtures/oci1-to-schema2.json
