+++ internal/image/fixtures/oci1.json
