+++ internal/image/fixtures/schema1-for-oci-config.json
