+++ internal/image/fixtures/schema1-to-oci1.json
