+++ internal/image/fixtures/schema1-to-schema2-config.json
