+++ internal/image/fixtures/schema1.json
