+++ internal/image/fixtures/schema2-all-media-types-to-oci1.json
