+++ internal/image/fixtures/schema2-all-media-types.json
