+++ internal/image/fixtures/schema2-config.json
