+++ internal/image/fixtures/schema2-invalid-media-type.json
