+++ internal/image/fixtures/schema2-to-oci1.json
