+++ internal/image/fixtures/schema2-to-schema1-by-docker.json
