+++ internal/image/fixtures/schema2.json
