+++ internal/image/fixtures/schema2list-variants.json
