+++ internal/image/manifest.go
