+++ internal/image/manifest_test.go
