+++ internal/image/memory.go
