+++ manifest/fixtures/schema2-to-schema1-by-docker.json
@@ -1,1 +1,1 @@
-../../image/fixtures/schema2-to-schema1-by-docker.json
+../../internal/image/fixtures/schema2-to-schema1-by-docker.json
